chore(githooks): Only check if staged files comply with formatting
Formatting and adding whole files in Git hooks makes using git add -p
impossible.  We would want the hook to apply fprettify on the staged
hunks only.  It might is possible to get the diff and apply it to the
index

fprettify --diff | git apply --cached

with the caveat that the diff generated by fprettify shows '-' as the
files and should be replace by the staged file name. sed to the rescue.