Suppress multiple module inclusions with Cray compiler
The Cray compiler warns when a module has been included multiple
times, e.g. when module B includes module A and module C includes
module A and B this will print a warning. Since there is seemingly
no harm in this, and in my experience not all compilers agree on
inclusion inheritance... this simply creates noise during
compilation.